requirements.txt
alembic==1.14.0
argcomplete==3.5.3
asgiref==3.7.2
awscli==1.37.6
bcrypt==4.2.1
blinker==1.9.0
boto==2.49.0
boto3==1.36.6
botocore==1.36.6
certifi==2023.7.22
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
Deprecated==1.2.17
Django==5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docutils==0.16
durationpy==0.9
et_xmlfile==2.0.0
fabric==3.2.2
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
hjson==3.1.0
idna==3.4
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kappa==0.6.0
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
placebo==0.9.0
prettytable==3.13.0
pyasn1==0.6.1
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.31.0
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.13.0
zappa==0.59.0
